pyproject.toml
[project]
name = "proyecto"
version = "0.1.0"
description = "Proyecto de IE0405 - Modelos Probabilísticos de Señales y Sistemas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "matplotlib>=3.10.7",
    "mkdocs-material>=9.6.22",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
